app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
pytest
debugpy
ruff
black
httpx
python-dotenv
prometheus-fastapi-instrumentator>=7.0.0
